fix debug warnings

+++ src/utility.cpp
@@ -8,12 +8,12 @@ void Utility::SaveWavefunction(const std::string& fname, const CVector& r, const
     std::ofstream file(fname);
     for (int i = 0; i < size; i++) {
         file << std::fixed << std::setprecision(14) << "#        r1         ";
-        for (int j = 0; j < wfn.size(); j++) {
+        for (size_t j = 0; j < wfn.size(); j++) {
             file << "     state" << (j < 10 ? "0" : "") << j << ".real         " << "state" << (j < 10 ? "0" : "") << j << ".imag    ";
         }
-        for (int j = 0; j < r.size(); j++) {
+        for (long int j = 0; j < r.size(); j++) {
             file << (!j ? "\n" : "") << std::setw(20) << r(j).real();
-            for (int k = 0; k < wfn.size(); k++) {
+            for (size_t k = 0; k < wfn.size(); k++) {
                 file << " " << std::setw(20) << wfn.at(k).at(std::min(i, (int)wfn.at(k).size() - 1))(j).real()
                 << " " << std::setw(20) << wfn.at(k).at(std::min(i, (int)wfn.at(k).size() - 1))(j).imag();
             }
